Initial attempt to add an L-NIC like shell into rocket-chip.
The goal is to be able to easily generate RocketCore systems with
or without LNIC interfaces.

The current code compiles fine and I've verified that it does not
break when LNIC is disabled. However, I needed to comment out all
of the LNIC related hardware generation code in the CSRFile because
it was causing None.get errors during elaboration for some reason.
So, enabling LNIC will definitely not work right now. Still need
to figure out what is causing that issue.