Initial implementation of L-NIC using GPRs.
Updated CSRFile to receive rx/tx commands from RegFile in RocketCore.
Updated IDecode to change register read don't cares to 0.
Added simple pkt generator to LNIC for testing.

NOTE: the hard part is dealing with pipeline flushes where the
discarded instructions need to "unread" the RxQueue. The current
implementation still does not deal with this correctly. However,
it does use the ctrl_killd signal to drive the RxQueue read enable,
so that instructions stalled in the decode stage will not read the
RxQueue multiple times.