Fied bug where Packetize enqueue logic was not properly setting num_init_pkts.
But now I realize that this is not the way to go.

Instead, the enqueue logic should only set toBtx state for pkts beyond
the first RTT_PKTS (i.e. pkts that are not sent immediately and need to
be pulled by credit events). Also, each bit should be set only when the
corresponding pkt is written.